RENAME : styledComponents.tsx=>common.tsx
